requirements.txt
# Core
pandas>=2.0
numpy>=1.25

# Machine learning & probabilistic models
scikit-learn>=1.3
hmmlearn>=0.3.0
scipy>=1.11

# Portfolio optimization
PyPortfolioOpt>=1.6.0
cvxpy>=1.3.2

# Backtesting & metrics
matplotlib>=3.8
seaborn>=0.12

# Trading APIs
yfinance>=0.2.33
alpaca-trade-api>=3.2.0
websockets>=10.0,<11.0

# Dashboard & visualization
streamlit>=1.26
streamlit-autorefresh>=0.1.2

# Misc
pyyaml>=6.0